+++ temba/flows/tests.py
@@ -1569,7 +1569,7 @@ class FlowTest(TembaTest):
 
         # reference to channel changed to match existing channel by name
         self.assertEqual(
-            {"uuid": str(channel.uuid), "name": "RapidPro Test"}, flow_def["nodes"][0]["actions"][4]["channel"],
+            {"uuid": str(channel.uuid), "name": "RapidPro Test"}, flow_def["nodes"][0]["actions"][4]["channel"]
         )
 
         # reference to ticketer unchanged because it matched existing ticketer by UUID
@@ -3841,7 +3841,7 @@ class ExportFlowResultsTest(TembaTest):
                 # make sure that we trigger logger
                 log_info_threshold.return_value = 1
 
-                with self.assertNumQueries(42):
+                with self.assertNumQueries(44):
                     workbook = self._export(flow, group_memberships=[devs])
 
                 self.assertEqual(len(captured_logger.output), 3)
@@ -4081,7 +4081,7 @@ class ExportFlowResultsTest(TembaTest):
         )
 
         # test without msgs or unresponded
-        with self.assertNumQueries(40):
+        with self.assertNumQueries(42):
             workbook = self._export(flow, include_msgs=False, responded_only=True, group_memberships=(devs,))
 
         tz = self.org.timezone
@@ -4147,7 +4147,7 @@ class ExportFlowResultsTest(TembaTest):
         )
 
         # test export with a contact field
-        with self.assertNumQueries(42):
+        with self.assertNumQueries(44):
             workbook = self._export(
                 flow,
                 include_msgs=False,
@@ -4377,7 +4377,7 @@ class ExportFlowResultsTest(TembaTest):
 
         contact1_run1, contact2_run1, contact3_run1, contact1_run2, contact2_run2 = FlowRun.objects.order_by("id")
 
-        with self.assertNumQueries(50):
+        with self.assertNumQueries(52):
             workbook = self._export(flow)
 
         tz = self.org.timezone
@@ -4647,7 +4647,7 @@ class ExportFlowResultsTest(TembaTest):
         )
 
         # test without msgs or unresponded
-        with self.assertNumQueries(33):
+        with self.assertNumQueries(35):
             workbook = self._export(flow, include_msgs=False, responded_only=True)
 
         tz = self.org.timezone
@@ -5637,6 +5637,12 @@ class FlowStartCRUDLTest(TembaTest, CRUDLTestMixin):
         self.assertContains(response, "all contacts")
         self.assertContains(response, "contacts who haven't already been through this flow")
         self.assertContains(response, "<b>1,234</b> runs")
+
+        response = self.assertListFetch(
+            list_url + "?type=manual", allow_viewers=True, allow_editors=True, context_objects=[start1]
+        )
+        self.assertTrue(response.context["filtered"])
+        self.assertEqual(response.context["url_params"], "?type=manual&")
 
 
 class AssetServerTest(TembaTest):