+++ temba/orgs/views.py
@@ -12,6 +12,7 @@ import pyotp
 import pytz
 import requests
 from packaging.version import Version
+from smartmin.users.views import Login
 from smartmin.views import (
     SmartCreateView,
     SmartCRUDL,
@@ -32,6 +33,7 @@ from django.contrib import messages
 from django.contrib.auth import authenticate, login, logout
 from django.contrib.auth.models import Group, User
 from django.contrib.auth.password_validation import validate_password
+from django.contrib.auth.views import LoginView as AuthLoginView
 from django.core.exceptions import ValidationError
 from django.core.validators import validate_email
 from django.db import IntegrityError
@@ -68,8 +70,11 @@ from temba.utils.http import http_headers
 from temba.utils.timezones import TimeZoneFormField
 from temba.utils.views import ComponentFormMixin, NonAtomicMixin
 
-from .models import BackupToken, Invitation, Org, OrgCache, OrgRole, TopUp, UserSettings, get_stripe_credentials
+from .models import BackupToken, Invitation, Org, OrgCache, OrgRole, TopUp, get_stripe_credentials
 from .tasks import apply_topups_task
+
+# session key for storing a two-factor enabled user's id once we've checked their password
+TWO_FACTOR_USER_SESSION_KEY = "_two_factor_user_id"
 
 
 def check_login(request):
@@ -253,7 +258,7 @@ class OrgSignupForm(forms.ModelForm):
     timezone = TimeZoneFormField(help_text=_("The timezone for your workspace"), widget=forms.widgets.HiddenInput())
 
     password = forms.CharField(
-        widget=InputWidget(attrs={"hide_label": True, "password": True, "placeholder": _("Password")},),
+        widget=InputWidget(attrs={"hide_label": True, "password": True, "placeholder": _("Password")}),
         validators=[validate_password],
         help_text=_("At least eight characters or more"),
     )
@@ -342,6 +347,102 @@ class OrgGrantForm(forms.ModelForm):
     class Meta:
         model = Org
         fields = "__all__"
+
+
+class LoginView(Login):
+    """
+    Overrides the smartmin login view to redirect users with 2FA enabled to a second verification view.
+    """
+
+    template_name = "orgs/login/login.haml"
+
+    def form_valid(self, form):
+        user = form.get_user()
+        if user.get_settings().two_factor_enabled:
+            self.request.session[TWO_FACTOR_USER_SESSION_KEY] = str(user.id)
+
+            verify_url = reverse("users.two_factor_verify")
+            redirect_url = self.get_redirect_url()
+            if redirect_url:
+                verify_url += f"?{self.redirect_field_name}={urlquote(redirect_url)}"
+
+            return HttpResponseRedirect(verify_url)
+
+        return super().form_valid(form)
+
+
+class BaseTwoFactorView(AuthLoginView):
+    def dispatch(self, request, *args, **kwargs):
+        # redirect back to login view if user hasn't completed that yet
+        user = self.get_user()
+        if not user:
+            return HttpResponseRedirect(reverse("users.login"))
+
+        return super().dispatch(request, *args, **kwargs)
+
+    def get_user(self):
+        user_id = self.request.session.get(TWO_FACTOR_USER_SESSION_KEY)
+        if user_id:
+            return User.objects.filter(id=user_id, is_active=True).first()
+        return None
+
+    def get_form_kwargs(self):
+        kwargs = super().get_form_kwargs()
+        kwargs["user"] = self.get_user()
+        return kwargs
+
+    def form_valid(self, form):
+        # set the user as actually authenticated now
+        login(self.request, self.get_user())
+
+        # remove our session key so if the user comes back this page they'll get directed to the login view
+        self.request.session.pop(TWO_FACTOR_USER_SESSION_KEY, None)
+
+        return HttpResponseRedirect(self.get_success_url())
+
+
+class TwoFactorVerifyView(BaseTwoFactorView):
+    """
+    View to let users with 2FA enabled verify their identity via an OTP from a device.
+    """
+
+    class Form(forms.Form):
+        otp = forms.CharField(max_length=6, required=True)
+
+        def __init__(self, request, user, *args, **kwargs):
+            self.user = user
+            super().__init__(*args, **kwargs)
+
+        def clean_otp(self):
+            data = self.cleaned_data["otp"]
+            if not self.user.verify_2fa(otp=data):
+                raise ValidationError(_("Incorrect OTP. Please try again."))
+            return data
+
+    form_class = Form
+    template_name = "orgs/login/two_factor_verify.haml"
+
+
+class TwoFactorBackupView(BaseTwoFactorView):
+    """
+    View to let users with 2FA enabled verify their identity using a backup token.
+    """
+
+    class Form(forms.Form):
+        token = forms.CharField(max_length=8, required=True)
+
+        def __init__(self, request, user, *args, **kwargs):
+            self.user = user
+            super().__init__(*args, **kwargs)
+
+        def clean_token(self):
+            data = self.cleaned_data["token"]
+            if not self.user.verify_2fa(backup_token=data):
+                raise ValidationError(_("Invalid backup token. Please try again."))
+            return data
+
+    form_class = Form
+    template_name = "orgs/login/two_factor_backup.haml"
 
 
 class UserCRUDL(SmartCRUDL):
@@ -501,48 +602,6 @@ class InferOrgMixin(object):
 
     def get_object(self, *args, **kwargs):
         return self.request.user.get_org()
-
-
-class PhoneRequiredForm(forms.ModelForm):
-    tel = forms.CharField(max_length=15, label="Phone Number", required=True)
-
-    def clean_tel(self):
-        if "tel" in self.cleaned_data:
-            tel = self.cleaned_data["tel"]
-            if not tel:  # pragma: needs cover
-                return tel
-
-            import phonenumbers
-
-            try:
-                normalized = phonenumbers.parse(tel, None)
-                if not phonenumbers.is_possible_number(normalized):  # pragma: needs cover
-                    raise forms.ValidationError(_("Invalid phone number, try again."))
-            except Exception:  # pragma: no cover
-                raise forms.ValidationError(_("Invalid phone number, try again."))
-            return phonenumbers.format_number(normalized, phonenumbers.PhoneNumberFormat.E164)
-
-    class Meta:
-        model = UserSettings
-        fields = ("tel",)
-
-
-class UserSettingsCRUDL(SmartCRUDL):
-    actions = ("update", "phone")
-    model = UserSettings
-
-    class Phone(ModalMixin, OrgPermsMixin, SmartUpdateView):
-        @classmethod
-        def derive_url_pattern(cls, path, action):
-            return r"^%s/%s/$" % (path, action)
-
-        def get_object(self, *args, **kwargs):
-            return self.request.user.get_settings()
-
-        fields = ("tel",)
-        form_class = PhoneRequiredForm
-        submit_button_name = _("Start Call")
-        success_url = "@orgs.usersettings_phone"
 
 
 class OrgCRUDL(SmartCRUDL):
@@ -1360,10 +1419,10 @@ class OrgCRUDL(SmartCRUDL):
     class ManageAccounts(InferOrgMixin, OrgPermsMixin, SmartUpdateView):
         class AccountsForm(forms.ModelForm):
             invite_emails = forms.CharField(
-                required=False, widget=InputWidget(attrs={"widget_only": True, "placeholder": _("Email Address")}),
+                required=False, widget=InputWidget(attrs={"widget_only": True, "placeholder": _("Email Address")})
             )
             invite_role = forms.ChoiceField(
-                choices=[], required=True, initial="V", label=_("Role"), widget=SelectWidget(),
+                choices=[], required=True, initial="V", label=_("Role"), widget=SelectWidget()
             )
 
             def __init__(self, org, *args, **kwargs):
@@ -1390,7 +1449,7 @@ class OrgCRUDL(SmartCRUDL):
                         widget=SelectWidget(),
                     )
                     remove_field = forms.BooleanField(
-                        required=False, label=" ", widget=CheckboxWidget(attrs={"widget_only": True}),
+                        required=False, label=" ", widget=CheckboxWidget(attrs={"widget_only": True})
                     )
 
                     self.fields.update(
@@ -1411,7 +1470,7 @@ class OrgCRUDL(SmartCRUDL):
                         disabled=True,
                     )
                     remove_field = forms.BooleanField(
-                        required=False, label=" ", widget=CheckboxWidget(attrs={"widget_only": True}),
+                        required=False, label=" ", widget=CheckboxWidget(attrs={"widget_only": True})
                     )
 
                     self.fields.update(
@@ -1486,9 +1545,9 @@ class OrgCRUDL(SmartCRUDL):
         def get_gear_links(self):
             links = []
             if self.request.user.get_org().id != self.get_object().id:
-                links.append(dict(title=_("Workspaces"), style="button-light", href=reverse("orgs.org_sub_orgs"),))
+                links.append(dict(title=_("Workspaces"), style="button-light", href=reverse("orgs.org_sub_orgs")))
 
-            links.append(dict(title=_("Home"), style="button-light", href=reverse("orgs.org_home"),))
+            links.append(dict(title=_("Home"), style="button-light", href=reverse("orgs.org_home")))
             return links
 
         def get_form_kwargs(self):
@@ -1564,31 +1623,26 @@ class OrgCRUDL(SmartCRUDL):
 
     class TwoFactor(ComponentFormMixin, InferOrgMixin, OrgPermsMixin, SmartFormView):
         class TwoFactorForm(forms.Form):
-            token = forms.CharField(
-                label=_("Authentication Token"),
-                help_text=_("Enter the code from your authentication application"),
+            otp = forms.CharField(
+                label=_("One-time Password (OTP)"),
+                help_text=_("Enter the one-time password from your authentication application."),
                 strip=True,
                 required=True,
             )
 
-            def __init__(self, *args, **kwargs):
-                self.request = kwargs.pop("request")
-                self.user_cache = None
+            def __init__(self, user, *args, **kwargs):
                 super().__init__(*args, **kwargs)
 
-            def clean_token(self):  # pragma: no cover
-                token = self.cleaned_data.get("token", None)
-                user_pk = self.request.user.pk
-                user = User.objects.get(pk=user_pk)
-                totp = pyotp.TOTP(user.get_settings().otp_secret)
-                token_valid = totp.verify(token, valid_window=2)
-                if not token_valid:
-                    raise forms.ValidationError(_("Invalid MFA token. Please try again."), code="invalid-token")
-                self.user_cache = user
-                return token
+                self.user = user
+
+            def clean_otp(self):
+                otp = self.cleaned_data["otp"]
+                if not self.user.verify_2fa(otp=otp):
+                    raise forms.ValidationError(_("Incorrect OTP. Please try again."))
+                return otp
 
         form_class = TwoFactorForm
-        fields = ("token",)
+        fields = ("otp",)
         success_url = "@orgs.org_two_factor"
         success_message = ""
         submit_button_name = _("Activate")
@@ -1596,78 +1650,40 @@ class OrgCRUDL(SmartCRUDL):
 
         def get_form_kwargs(self):
             kwargs = super().get_form_kwargs()
-            kwargs["request"] = self.request
+            kwargs["user"] = self.request.user
             return kwargs
 
-        def get(self, request, *args, **kwargs):
-            user = self.request.user
-            form = self.get_form()
-            secret = pyotp.random_base32()
-
-            user_settings = user.get_settings()
-            user_settings.otp_secret = secret
-            user_settings.save()
-            secret_url = self.get_secret_url()
-            return self.render_to_response(self.get_context_data(form=form, secret_url=secret_url))
-
         def post(self, request, *args, **kwargs):
+            user = self.get_user()
             form = self.get_form()
-            if "disable_two_factor_auth" in request.POST:
-                self.disable_two_factor_auth()
-            if "get_backup_tokens" in request.POST:
-                tokens = self.get_backup_tokens()
-                data = {"tokens": tokens}
-                return JsonResponse(data)
-            elif "generate_backup_tokens" in request.POST:
-                tokens = self.generate_backup_tokens()
-                data = {"tokens": tokens}
-                return JsonResponse(data)
+            action = request.POST.get("action", "")
+
+            if action == "disable":
+                user.disable_2fa()
+            elif action == "regenerate_backup_tokens":
+                BackupToken.generate_for_user(user)
+                return JsonResponse({"tokens": self.get_backup_tokens(user)})
             elif form.is_valid():
-                self.generate_backup_tokens()
-                user = self.request.user
-                user_settings = user.get_settings()
-                user_settings.two_factor_enabled = True
-                user_settings.save()
-            secret_url = self.get_secret_url()
-            return self.render_to_response(self.get_context_data(form=form, secret_url=secret_url))
+                user.enable_2fa()
+
+            return self.render_to_response(self.get_context_data(form=form))
 
         def get_context_data(self, **kwargs):
             context = super().get_context_data(**kwargs)
+
+            brand = self.request.branding["name"]
             user = self.get_user()
             user_settings = user.get_settings()
-            context["user_settings"] = user_settings
+            otp_secret = user_settings.otp_secret
+            secret_url = pyotp.TOTP(otp_secret).provisioning_uri(user.username, issuer_name=brand)
+
+            context["two_factor_enabled"] = user_settings.two_factor_enabled
+            context["secret_url"] = secret_url
+            context["backup_tokens"] = self.get_backup_tokens(user)
             return context
 
-        def get_secret_url(self):
-            user = self.request.user
-            otp_secret = user.get_settings().otp_secret
-            if otp_secret:
-                secret_url = pyotp.TOTP(otp_secret).provisioning_uri(user.username, issuer_name="Rapidpro")
-            return secret_url
-
-        def disable_two_factor_auth(self):
-            self.delete_backup_tokens()
-            user = self.get_user()
-            user_settings = user.get_settings()
-            user_settings.two_factor_enabled = False
-            user_settings.save()
-
-        def generate_backup_tokens(self):
-            user = self.get_user()
-            self.delete_backup_tokens()
-            for backup in range(10):
-                BackupToken.objects.create(settings=user.get_settings(), created_by=user, modified_by=user)
-            tokens = [backup.token for backup in BackupToken.objects.filter(settings__user=user)]
-            return tokens
-
-        def get_backup_tokens(self):
-            user = self.get_user()
-            tokens = [backup.token for backup in BackupToken.objects.filter(settings__user=user)]
-            return tokens
-
-        def delete_backup_tokens(self):
-            user = self.get_user()
-            BackupToken.objects.filter(settings__user=user).delete()
+        def get_backup_tokens(self, user):
+            return [{"token": t.token, "is_used": t.is_used} for t in user.backup_tokens.all()]
 
     class Service(SmartFormView):
         class ServiceForm(forms.Form):
@@ -2491,7 +2507,7 @@ class OrgCRUDL(SmartCRUDL):
                 if len(links) > 0:
                     links.append(dict(divider=True))
 
-                links.append(dict(title=_("Help"), href=settings.HELP_URL,))
+                links.append(dict(title=_("Help"), href=settings.HELP_URL))
 
             if len(links) > 0:
                 links.append(dict(divider=True))
@@ -2815,7 +2831,7 @@ class OrgCRUDL(SmartCRUDL):
         class OrgForm(forms.ModelForm):
             name = forms.CharField(max_length=128, label=_("Workspace Name"), help_text="", widget=InputWidget())
             timezone = TimeZoneFormField(
-                label=_("Timezone"), help_text="", widget=SelectWidget(attrs={"searchable": True}),
+                label=_("Timezone"), help_text="", widget=SelectWidget(attrs={"searchable": True})
             )
 
             class Meta: